pyproject.toml
[tool.poetry]
name = "plant-doctor"
version = "0.1.0"
description = ""
authors = ["bhanpuramufaddal <bhanpuramufaddal@gmail.com>"]
readme = "README.md"
packages = [{include = "plant_doctor"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.27.1"
pydantic = "^2.6.2"
pillow = "^10.2.0"
geopy = "^2.4.1"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.109.2"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
